feat: add types for Breadcrumbs
