+++ .all-contributorsrc
@@ -1278,6 +1278,15 @@
         "code",
         "doc"
       ]
+    },
+    {
+      "login": "Tresau-IBM",
+      "name": "Tresau-IBM",
+      "avatar_url": "https://avatars.githubusercontent.com/u/148357638?v=4",
+      "profile": "https://github.com/Tresau-IBM",
+      "contributions": [
+        "code"
+      ]
     }
   ],
   "commitConvention": "none"
